+++ src/buildgen/common/__init__.py
@@ -4,6 +4,7 @@ from buildgen.common.utils import UniqueList, check_output, env_var, always_true
 from buildgen.common.platform import PLATFORM, PythonSystem
 from buildgen.common.base import BaseGenerator, BaseBuilder
 from buildgen.common.project import ProjectConfig, TargetConfig, DependencyConfig
+from buildgen.common.config import UserConfig, load_user_config
 
 __all__ = [
     "UniqueList",
@@ -17,4 +18,6 @@ __all__ = [
     "ProjectConfig",
     "TargetConfig",
     "DependencyConfig",
+    "UserConfig",
+    "load_user_config",
 ]